backend/requirements.txt
fastapi
uvicorn==0.24.0
websockets==12.0
pydantic
numpy==1.24.3
tensorflow
scikit-learn==1.3.2
python-multipart==0.0.6
pandas==2.1.4
plotly
streamlit